+++ src/Checkers/AbstractChecker.php
@@ -71,16 +71,6 @@ abstract class AbstractChecker
         return array_reverse($this->fixed);
     }
 
-    protected function currentVersion(): string
-    {
-        preg_match('/\d+\.\d+\.\d+/', $this->file->getContents(), $matches);
-        var_dump($matches);
-        die();
-        if ($matches) {
-            return $matches;
-        }
-    }
-
     protected function isVersionLine(string $line): bool
     {
         return substr($line, 0, 3) === '## ';