+++ src/Checkers/IncreasedChecker.php
@@ -43,18 +43,72 @@ class IncreasedChecker extends AbstractChecker
             return;
         }
 
-        file_put_contents($this->checksumFile, md5($this->file->getContents()));
+        $this->markChangedChangelog();
     }
 
     protected function fix(): void
     {
-        $commits = shell_exec(sprintf('git commit --oneline -n %n'), $this->options->fail_after);
+        if ($this->changelogChanged()) {
+            return;
+        }
+        $committedTitle = '### Committed';
+        $failAfterXCommits = $this->options->fail_after;
+        $command = sprintf('git log --oneline -n %d', $failAfterXCommits);
+        $commits = shell_exec($command);
+        $lines = array_filter(explode(PHP_EOL, $commits));
+        $newestCommit = $lines[0] ?? '';
+        $commitLog = $this->getCommitLog();
+        $position = strpos($commitLog, $newestCommit);
+        if ($position > 0) {
+            return;
+        }
         $commitArray = explode(PHP_EOL, $commits);
+        $commitArray = array_filter($commitArray);
+        $commitArray = array_map(fn ($commit) => "- $commit", $commitArray);
+        $hasFixedCommits = strpos($this->file->getContents(), $committedTitle) !== false;
+        if (!empty($commitArray) && !$hasFixedCommits) {
+            array_unshift($commitArray, $committedTitle);
+        }
+
         foreach ($this->file as $line) {
+            if ($hasFixedCommits) {
+                if (strpos($line, $committedTitle) !== false) {
+                    $lineIndex = $this->file->key() + 1;
+                    break;
+                }
+                continue;
+            }
             if (strpos($line, '[Unreleased]')) {
-                $this->file->includeLinesAfter($commitArray, $this->file->key());
+                $lineIndex = $this->file->key() + 2;
+                break;
             }
         }
+        if (!isset($lineIndex)) {
+            $this->addErrorMessage(
+                sprintf(
+                    '%s could not be added to the Changelog, because the [Unreleased] tag is missing',
+                    $commits
+                )
+            );
+            return;
+        }
+        $message = sprintf(
+            'Added commit changes "%s" to line %s',
+            $lineIndex,
+            $this->file->lineNumber()
+        );
+        $this->addFixedMessage($message);
+        foreach ($commitArray as $key => $commit) {
+            if (strpos($this->file->getContents(), $commit)) {
+                unset($commitArray[$key]);
+            }
+        }
+        $commitArray = array_filter($commitArray);
+        if (!empty($commitArray)) {
+            $this->file->includeLinesAfter($commitArray, $lineIndex);
+        }
+        file_put_contents($this->commitLogFile, implode(PHP_EOL, $lines));
+        $this->changelogChanged();
     }
 
 
@@ -79,5 +133,10 @@ class IncreasedChecker extends AbstractChecker
         $currentChecksum = md5($this->file->getContents());
         $lastChecksum = $this->getChecksum();
         return $lastChecksum !== $currentChecksum;
+    }
+
+    protected function markChangedChangelog(): void
+    {
+        file_put_contents($this->checksumFile, md5($this->file->getContents()));
     }
 }